Crate matrix_sdk

A high-level, batteries-included Matrix client library written in Rust.

This crate seeks to be a general-purpose library for writing software using the Matrix Client-Server API to communicate with a Matrix homeserver. If you’re writing a typical Matrix client or bot, this is likely the crate you need.

However, the crate is designed in a modular way and depends on several other lower-level crates. If you’re attempting something more custom, you might be interested in these:

matrix_sdk_base: A no-network-IO client state machine which can be used to embed a Matrix client into an existing network stack or to build a new Matrix client library on top.

matrix_sdk_crypto: A no-network-IO encryption state machine which can be used to add Matrix E2EE support into an existing client or library.

Getting started

The central component you’ll be interacting with is the Client. A basic use case will include instantiating the client, logging in as a user, registering some event handlers and then syncing.

This is demonstrated in the example below.

use matrix_sdk::{
    Client, config::SyncSettings,
    ruma::{user_id, events::room::message::SyncRoomMessageEvent},
};

#[tokio::main]
async fn main() -> anyhow::Result<()> {
    let alice = user_id!("@alice:example.org");
    let client = Client::builder().server_name(alice.server_name()).build().await?;

    // First we need to log in.
    client.login_username(alice, "password").send().await?;

    client.add_event_handler(|ev: SyncRoomMessageEvent| async move {
        println!("Received a message {:?}", ev);
    });

    // Syncing is important to synchronize the client state with the server.
    // This method will never return unless there is an error.
    client.sync(SyncSettings::default()).await?;

    Ok(())
}

More examples can be found in the examples directory.

Crate Feature Flags

The following crate feature flags are available:

Feature	Default	Description

anyhow	No	Better logging for event handlers that return anyhow::Result

e2e-encryption	Yes	End-to-end encryption (E2EE) support

eyre	No	Better logging for event handlers that return eyre::Result

image-proc	No	Image processing for generating thumbnails

image-rayon	No	Enables faster image processing

js	No	Enables JavaScript API usage for things like the current system time on WASM (does nothing on other targets)

markdown	No	Support for sending Markdown-formatted messages

qrcode	Yes	QR code verification support

sled	Yes	Persistent storage of state and E2EE data (optionally, if feature e2e-encryption is enabled), via Sled

indexeddb	No	Persistent storage of state and E2EE data (optionally, if feature e2e-encryption is enabled) for browsers, via IndexedDB

socks	No	SOCKS support in the default HTTP client, reqwest

sso-login	No	Support for SSO login with a local HTTP server

Enabling logging

Users of the matrix-sdk crate can enable log output by depending on the tracing-subscriber crate and including the following line in their application (e.g. at the start of main):

tracing_subscriber::fmt::init();

The log output is controlled via the RUST_LOG environment variable by setting it to one of the error, warn, info, debug or trace levels. The output is printed to stdout.

The RUST_LOG variable also supports a more advanced syntax for filtering log output more precisely, for instance with crate-level granularity. For more information on this, check out the tracing_subscriber documentation.

Re-exports

pub use bytes;

pub use reqwest;

pub use ruma;

pub use media::Media;

Modules

attachment

Types and traits for attachments.

config

Configuration to change the behaviour of the Client.

deserialized_responses

SDK-specific variations of response types from Ruma.

encryptione2e-encryption

End-to-end encryption related types

event_handler

Types and traits related for event handlers. For usage, see Client::add_event_handler.

executor

Abstraction over an executor so we can spawn tasks under WASM the same way we do usually.

instant

media

High-level media API.

room

High-level room API

sync

The SDK’s representation of the result of a /sync request.

timeout

Structs

Account

A high-level API to manage the client owner’s account.

BaseRoom

The underlying room data structure collecting state for joined, left and invited rooms.

BaseRoomMember

A member of a room.

Client

An async/await enabled Matrix client.

ClientBuilder

Builder that allows creating and configuring various parts of a Client.

LoginBuilder

Builder type used to configure optional settings for logging in with a username or token.

OwnedServerName

Owned variant of ServerName

RoomInfo

The underlying pure data structure for joined and left rooms.

ServerName

A Matrix-spec compliant server name.

Session

A user session, containing an access token, an optional refresh token and information about the associated user account.

SsoLoginBuildersso-login

Builder type used to configure optional settings for logging in via SSO.

StateChanges

Store state changes and pass them to the StateStore.

UnknownToken

Wrapper struct for ErrorKind::UnknownToken

Enums

ClientBuildError

Errors that can happen in ClientBuilder::build.

DisplayName

The name of the room, either from the metadata or calculated according to matrix specification

Error

Internal representation of errors.

HttpError

An HTTP error, representing either a connection error or an error while converting the raw HTTP response into a Matrix response.

IdParseError

An error encountered when trying to parse an invalid ID string.

ImageErrorimage-proc

All possible errors that can happen during image processing.

LoopCtrl

Enum controlling if a loop running callbacks should continue or abort.

RefreshTokenError

Errors that can happen when refreshing an access token.

RoomState

Enum keeping track in which state the room is, e.g. if our own user is joined, invited, or has left the room.

RumaApiError

An error response from a Matrix API call, using a client API specific representation if the endpoint is from that.

StoreError

State store specific error type.

Traits

AsyncTraitDeps

Super trait that is used for our store traits, this trait will differ if it’s used on WASM. WASM targets will not require Send and Sync to have implemented, while other targets will.

HttpSend

Abstraction around the http layer. The allows implementors to use different http libraries.

SendOutsideWasmNon-WebAssembly

Alias for Send on non-wasm, empty trait (implemented by everything) on wasm.

SyncOutsideWasmNon-WebAssembly

Alias for Sync on non-wasm, empty trait (implemented by everything) on wasm.

Functions

sanitize_server_name

Creates a server name from a user supplied string. The string is first sanitized by removing whitespace, the http(s) scheme and any trailing slashes before being parsed.

Type Definitions

HttpResult

Result type of a pure HTTP request.

Result

Result type of the matrix-sdk.

Attribute Macros

async_trait