Module matrix_sdk::api::r0::membership

Endpoints for room membership.

Modules

ban_user
POST /_matrix/client/r0/rooms/{roomId}/ban

forget_room
POST /_matrix/client/r0/rooms/{roomId}/forget

get_member_events
GET /_matrix/client/r0/rooms/{roomId}/members

invite_user
POST /_matrix/client/r0/rooms/{roomId}/invite

join_room_by_id
POST /_matrix/client/r0/rooms/{roomId}/join

join_room_by_id_or_alias
POST /_matrix/client/r0/join/{roomIdOrAlias}

joined_members
GET /_matrix/client/r0/rooms/{roomId}/joined_members

joined_rooms
GET /_matrix/client/r0/joined_rooms

kick_user
POST /_matrix/client/r0/rooms/{roomId}/kick

leave_room
POST /_matrix/client/r0/rooms/{roomId}/leave

unban_user
POST /_matrix/client/r0/rooms/{roomId}/unban

Structs

IncomingInvite3pid
‘Incoming’ variant of Invite3pid.

IncomingThirdPartySigned
‘Incoming’ variant of ThirdPartySigned.

Invite3pid
Represents third party IDs to invite to the room.

ThirdPartySigned
A signature of an m.third_party_invite token to prove that this user owns a third party identity which has been invited to the room.