Module matrix_sdk::api::r0::uiaa

Module for User-Interactive Authentication API types.

Structs

AuthFlow
Description of steps required to authenticate via the User-Interactive Authentication API.

UiaaInfo
Information about available authentication flows and status for User-Interactive Authenticiation API.

Enums

AuthData
Additional authentication information for the user-interactive authentication API.

IncomingAuthData
‘Incoming’ variant of AuthData.

UiaaResponse
Contains either a User-Interactive Authentication API response body or a Matrix error.