Function matrix_sdk::attachment::generate_image_thumbnail

pub fn generate_image_thumbnail<R: BufRead + Seek>(
    content_type: &Mime,
    reader: R,
    size: Option<(u32, u32)>
) -> Result<(Vec<u8>, BaseThumbnailInfo), ImageError>

Available on crate feature image-proc only.

Generate a thumbnail for an image.

This is a convenience method that uses the image crate.

Arguments

content_type - The type of the media, this will be used as the content-type header.

reader - A Reader that will be used to fetch the raw bytes of the media.

size - The size of the thumbnail in pixels as a (width, height) tuple. If set to None, defaults to (800, 600).

Examples

let path = PathBuf::from("/home/example/my-cat.jpg");
let mut image = BufReader::new(File::open(path)?);

let (thumbnail_data, thumbnail_info) = generate_image_thumbnail(
    &mime::IMAGE_JPEG,
    &mut image,
    None
)?;
let mut cursor = Cursor::new(thumbnail_data);
let config = AttachmentConfig::with_thumbnail(Thumbnail {
    reader: &mut cursor,
    content_type: &mime::IMAGE_JPEG,
    info: Some(thumbnail_info),
});

image.rewind()?;

if let Some(room) = client.get_joined_room(&room_id) {
    room.send_attachment(
        "My favorite cat",
        &mime::IMAGE_JPEG,
        &mut image,
        config,
    ).await?;
}