Struct matrix_sdk::attachment::AttachmentConfig

pub struct AttachmentConfig { /* private fields */ }

Configuration for sending an attachment.

Implementations

impl AttachmentConfig

pub fn new() -> Self

Create a new default AttachmentConfig without providing a thumbnail.

To provide a thumbnail use AttachmentConfig::with_thumbnail().

pub fn generate_thumbnail(self, size: Option<(u32, u32)>) -> Self

Available on crate feature image-proc only.

Generate the thumbnail to send for this media.

Uses generate_image_thumbnail().

Thumbnails can only be generated for supported image attachments. For more information, see the image crate.

Arguments

size - The size of the thumbnail in pixels as a (width, height) tuple. If set to None, defaults to (800, 600).

pub fn with_thumbnail(thumbnail: Thumbnail) -> Self

Create a new default AttachmentConfig with a thumbnail.

Arguments

thumbnail - The thumbnail of the media. If the content_type does not support it (eg audio clips), it is ignored.

To generate automatically a thumbnail from an image, use AttachmentConfig::new() and AttachmentConfig::generate_thumbnail().

pub fn txn_id(self, txn_id: &TransactionId) -> Self

Set the transaction ID to send.

Arguments

txn_id - A unique ID that can be attached to a MessageEvent held in its unsigned field as transaction_id. If not given, one is created for the message.

pub fn info(self, info: AttachmentInfo) -> Self

Set the media metadata to send.

Arguments

info - The metadata of the media. If the AttachmentInfo type doesn’t match the content_type, it is ignored.