Module matrix_sdk::encryption::verification

Available on crate feature e2e-encryption only.

Interactive verification for E2EE capable users and devices in Matrix.

The SDK supports interactive verification of devices and users, this module contains types that model and support different verification flows.

A verification flow usually starts its life as a VerificationRequest, the request can then be accepted, or it needs to be accepted by the other side of the verification flow.

Once both sides have agreed to pereform the verification, and the VerificationRequest::is_ready() method returns true, the verification can transition into one of the supported verification flows:

SasVerification - Interactive verification using a short authentication string.

QrVerification - Interactive verification using QR codes.

Structs

AcceptSettings

Customize the accept-reply for a verification process

AcceptedProtocols

Struct containing the protocols that were agreed to be used for the SAS flow.

CancelInfo

Information about the cancellation of a verification request or verification flow.

Emoji

An emoji that is used for interactive verification using a short auth string.

EmojiShortAuthString

The short auth string for the emoji method of SAS verification.

QrVerificationqrcode

An object controlling QR code style key verification flows.

SasVerification

An object controlling the short auth string verification flow.

VerificationRequest

An object controlling the interactive verification flow.

Enums

DecodingErrorqrcode

Error type describing errors that happen while QR data is being decoded.

EncodingErrorqrcode

Error type describing errors that happen while QR data is being encoded.

QrVerificationDataqrcode

An enum representing the different modes a QR verification can be in.

SasState

An Enum describing the state the SAS verification is in.

ScanErrorqrcode

An error for the different failure modes that can happen during the validation of a scanned QR code.

Verification

An enum over the different verification types the SDK supports.

Functions

format_emojis

Format the the list of emojis as a two line string.