Struct matrix_sdk::encryption::verification::Emoji

pub struct Emoji {
    pub symbol: &'static str,
    pub description: &'static str,
}

An emoji that is used for interactive verification using a short auth string.

This will contain a single emoji and description from the list of emojis from the spec.

Fields
symbol: &'static str

The emoji symbol that represents a part of the short auth string, for example: ๐ถ

description: &'static str

The description of the emoji, for example โDogโ.