Struct matrix_sdk::encryption::verification::VerificationRequest

pub struct VerificationRequest { /* private fields */ }

Available on crate feature e2e-encryption only.

An object controlling the interactive verification flow.

Implementations

impl VerificationRequest

pub fn is_done(&self) -> bool

Has this verification finished.

pub fn is_cancelled(&self) -> bool

Has the verification been cancelled.

pub fn flow_id(&self) -> &str

Get the transaction id of this verification request

pub fn cancel_info(&self) -> Option<CancelInfo>

Get info about the cancellation if the verification request has been cancelled.

pub fn own_user_id(&self) -> &UserId

Get our own user id.

pub fn is_passive(&self) -> bool

Has the verification request been answered by another device.

pub fn is_ready(&self) -> bool

Is the verification request ready to start a verification flow.

pub fn we_started(&self) -> bool

Did we initiate the verification flow.

pub fn other_user_id(&self) -> &UserId

Get the user id of the other user participating in this verification flow.

pub fn is_self_verification(&self) -> bool

Is this a verification that is veryfying one of our own devices.

pub fn their_supported_methods(&self) -> Option<Vec<VerificationMethod>>

Get the supported verification methods of the other side.

Will be present only if the other side requested the verification or if we’re in the ready state.

pub async fn accept(&self) -> Result<()>

Accept the verification request.

This method will accept the request and signal by default that it supports the m.sas.v1, the m.qr_code.show.v1, and m.reciprocate.v1 method. If the qrcode feature is disabled it will only signal that it supports the m.sas.v1 method.

If QR code scanning should be supported or QR code showing shouldn’t be supported the accept_with_methods() method should be used instead.

pub async fn accept_with_methods( &self, methods: Vec<VerificationMethod> ) -> Result<()>

Accept the verification request signaling that our client supports the given verification methods.

Arguments

methods - The methods that we should advertise as supported by us.

pub async fn generate_qr_code(&self) -> Result<Option<QrVerification>>

Available on crate feature qrcode only.

Generate a QR code

pub async fn scan_qr_code( &self, data: QrVerificationData ) -> Result<Option<QrVerification>>

Available on crate feature qrcode only.

Start a QR code verification by providing a scanned QR code for this verification flow.

Returns an Error if the QR code isn’t valid or sending a reciprocate event to the other side fails, None if the verification request isn’t in the ready state or we don’t support QR code verification, otherwise a newly created QrVerification object which will be used for the remainder of the verification flow.

pub async fn start_sas(&self) -> Result<Option<SasVerification>>

Transition from this verification request into a SAS verification flow.

pub async fn cancel(&self) -> Result<()>

Cancel the verification request