Enum matrix_sdk::Error

#[non_exhaustive]
pub enum Error {

    Http(HttpError),
    AuthenticationRequired,
    InsufficientData,
    BadCryptoStoreState,
    NoOlmMachine,
    SerdeJson(JsonError),
    Io(IoError),
    CryptoStoreError(CryptoStoreError),
    OlmError(OlmError),
    MegolmError(MegolmError),
    DecryptorError(DecryptorError),
    StateStore(StoreError),
    Identifier(IdParseError),
    Url(UrlParseError),
    QrCodeScanError(ScanError),
    UserTagName(InvalidUserTagName),
    ImageError(ImageError),
    InconsistentState,
    UnknownError(Box<dyn Error + Send + Sync>),

}

Internal representation of errors.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Http(HttpError)

Error doing an HTTP request.

AuthenticationRequired

Queried endpoint requires authentication but was called on an anonymous client.

InsufficientData

This request failed because the local data wasnโt sufficient.

BadCryptoStoreState

Available on crate feature e2e-encryption only.

Attempting to restore a session after the olm-machine has already been set up fails

NoOlmMachine

Available on crate feature e2e-encryption only.

Attempting to access the olm-machine but it is not yet available.

SerdeJson(JsonError)

An error de/serializing type for the StateStore

Io(IoError)

An IO error happened.

CryptoStoreError(CryptoStoreError)

Available on crate feature e2e-encryption only.

An error occurred in the crypto store.

OlmError(OlmError)

Available on crate feature e2e-encryption only.

An error occurred during a E2EE operation.

MegolmError(MegolmError)

Available on crate feature e2e-encryption only.

An error occurred during a E2EE group operation.

DecryptorError(DecryptorError)

Available on crate feature e2e-encryption only.

An error occurred during decryption.

StateStore(StoreError)

An error occurred in the state store.

Identifier(IdParseError)

An error encountered when trying to parse an identifier.

Url(UrlParseError)

An error encountered when trying to parse a url.

QrCodeScanError(ScanError)

Available on crate feature qrcode only.

An error while scanning a QR code.

UserTagName(InvalidUserTagName)

An error encountered when trying to parse a user tag name.

ImageError(ImageError)

Available on crate feature image-proc only.

An error while processing images.

InconsistentState

The client is in inconsistent state. This happens when we set a room to a specific type, but then cannot get it in this type.

UnknownError(Box<dyn Error + Send + Sync>)

An other error was raised this might happen because encryption was enabled on the base-crate but not here and that raised.

Implementations

impl Error

pub fn as_ruma_api_error(&self) -> Option<&RumaApiError>

If self is Http(Api(Server(e))), returns Some(e).

Otherwise, returns None.

pub fn as_client_api_error(&self) -> Option<&Error>

Shorthand for .as_ruma_api_error().and_then(RumaApiError::as_client_api_error).

pub fn client_api_error_kind(&self) -> Option<&ErrorKind>

If self is a server error in the errcode + error format expected for client-API endpoints, returns the error kind (errcode).

pub fn as_uiaa_response(&self) -> Option<&UiaaInfo>

Try to destructure the error into an universal interactive auth info.

Some requests require universal interactive auth, doing such a request will always fail the first time with a 401 status code, the response body will contain info how the client can authenticate.

The request will need to be retried, this time containing additional authentication data.

This method is an convenience method to get to the info the server returned on the first, failed request.