Enum matrix_sdk::HttpError

pub enum HttpError {
    Reqwest(ReqwestError),
    AuthenticationRequired,
    NotClientRequest,
    Api(FromHttpResponseError<RumaApiError>),
    IntoHttp(IntoHttpError),
    UnableToCloneRequest,
    RefreshToken(RefreshTokenError),
}

An HTTP error, representing either a connection error or an error while converting the raw HTTP response into a Matrix response.

Variants

Reqwest(ReqwestError)

An error at the HTTP layer.

AuthenticationRequired

Queried endpoint requires authentication but was called on an anonymous client.

NotClientRequest

Queried endpoint is not meant for clients.

Api(FromHttpResponseError<RumaApiError>)

An error converting between ruma_*_api types and Hyper types.

IntoHttp(IntoHttpError)

An error converting between ruma_client_api types and Hyper types.

UnableToCloneRequest

The given request canโt be cloned and thus canโt be retried.

RefreshToken(RefreshTokenError)

An error occurred while refreshing the access token.

Implementations

impl HttpError

pub fn as_ruma_api_error(&self) -> Option<&RumaApiError>

If self is Api(Server(e)), returns Some(e).

Otherwise, returns None.

pub fn as_client_api_error(&self) -> Option<&Error>

Shorthand for .as_ruma_api_error().and_then(RumaApiError::as_client_api_error).

pub fn client_api_error_kind(&self) -> Option<&ErrorKind>

If self is a server error in the errcode + error format expected for client-API endpoints, returns the error kind (errcode).

pub fn as_uiaa_response(&self) -> Option<&UiaaInfo>

Try to destructure the error into an universal interactive auth info.

Some requests require universal interactive auth, doing such a request will always fail the first time with a 401 status code, the response body will contain info how the client can authenticate.

The request will need to be retried, this time containing additional authentication data.

This method is an convenience method to get to the info the server returned on the first, failed request.