Trait matrix_sdk::event_handler::EventHandlerResult

pub trait EventHandlerResult: Sized { }

Return types supported for event handlers implement this trait.

It is not meant to be implemented outside of matrix-sdk.