Module matrix_sdk::events::key::verification

Modules for events in the m.key.verification namespace.

This module also contains types shared by events in its child namespaces.

The MSC for the in-room variants of the m.key.verification.* events can be found here.

Modules

accept
Types for the m.key.verification.accept event.

cancel
Types for the m.key.verification.cancel event.

done
Types for the m.key.verification.done event.

key
Types for the m.key.verification.key event.

mac
Types for the m.key.verification.mac event.

ready
Types for the m.key.verification.ready event.

request
Types for the m.key.verification.request event.

start
Types for the m.key.verification.start event.

Structs

Relation
The relation that contains info which event the reaction is applying to.

Enums

HashAlgorithm
A hash algorithm.

KeyAgreementProtocol
A key agreement protocol.

MessageAuthenticationCode
A message authentication code algorithm.

ShortAuthenticationString
A Short Authentication String method.

VerificationMethod
A Short Authentication String (SAS) verification method.