Module matrix_sdk::events::macros

Re-export of all the derives needed to create your own event types.

Derive Macros

BasicEventContent
Generates an implementation of ruma_events::BasicEventContent and it’s super traits.

EphemeralRoomEventContent
Generates an implementation of ruma_events::EphemeralRoomEventContent and it’s super traits.

Event
Generates implementations needed to serialize and deserialize Matrix events.

MessageEventContent
Generates an implementation of ruma_events::MessageEventContent and it’s super traits.

StateEventContent
Generates an implementation of ruma_events::StateEventContent and it’s super traits.