Module matrix_sdk::events::room::message

Types for the m.room.message event.

Modules

feedback
Types for the m.room.message.feedback event.

Structs

AudioInfo
Metadata about an audio clip.

AudioMessageEventContent
The payload for an audio message.

EmoteMessageEventContent
The payload for an emote message.

FileInfo
Metadata about a file.

FileMessageEventContent
The payload for a file message.

FormattedBody
Common message event content fields for message types that have separate plain-text and formatted representations.

ImageMessageEventContent
The payload for an image message.

InReplyTo
Information about the event a “rich reply” is replying to.

KeyVerificationRequestEventContent
The payload for a key verification request message.

LocationInfo
Thumbnail info associated with a location.

LocationMessageEventContent
The payload for a location message.

MessageEventContent
The payload for MessageEvent.

NoticeMessageEventContent
The payload for a notice message.

RedactedMessageEventContent
The payload for a redacted MessageEventContent

ServerNoticeMessageEventContent
The payload for a server notice message.

TextMessageEventContent
The payload for a text message.

VideoInfo
Metadata about a video.

VideoMessageEventContent
The payload for a video message.

Enums

LimitType
Types of usage limits.

MessageFormat
The format for the formatted representation of a message body.

MessageType
The content that is specific to each message type variant.

Relation
Enum modeling the different ways relationships can be expressed in a m.relates_to field of an m.room.message event.

ServerNoticeType
Types of server notices.

Type Definitions

MessageEvent
This event is used when sending messages in a room.