Module matrix_sdk::media

High-level media API.

Structs

Media

A high-level API to interact with the media API.

MediaFileHandle

A file handle that takes ownership of a media file on disk. When the handle is dropped, the file will be removed from the disk.

MediaRequest

A request for media data.

MediaThumbnailSize

The requested size of a media thumbnail.

PersistError

Error returned when MediaFileHandle::persist fails.

Enums

MediaFormat

The requested format of a media file.

Traits

MediaEventContent

Trait for media event content.

UniqueKey

A trait to uniquely identify values of the same type.

Type Aliases

SendUploadRequest

IntoFuture returned by Media::upload.