Struct matrix_sdk::room::Common

pub struct Common { /* private fields */ }

A struct containing methods that are common for Joined, Invited and Left Rooms

Implementations

impl Common

pub fn new(client: Client, room: BaseRoom) -> Self

Create a new room::Common

Arguments

client - The client used to make requests.

room - The underlying room.

pub async fn avatar(&self, format: MediaFormat) -> Result<Option<Vec<u8>>>

Gets the avatar of this room, if set.

Returns the avatar. If a thumbnail is requested no guarantee on the size of the image is given.

Arguments

format - The desired format of the avatar.

Example

let client = Client::new(homeserver).await.unwrap();
client.login(user, "password", None, None).await.unwrap();
let room_id = room_id!("!roomid:example.com");
let room = client
    .get_joined_room(&room_id)
    .unwrap();
if let Some(avatar) = room.avatar(MediaFormat::File).await.unwrap() {
    std::fs::write("avatar.png", avatar);
}

pub async fn messages(&self, options: MessagesOptions<'_>) -> Result<Messages>

Sends a request to /_matrix/client/r0/rooms/{room_id}/messages and returns a Messages struct that contains a chunk of room and state events (RoomEvent and AnyStateEvent).

With the encryption feature, messages are decrypted if possible. If decryption fails for an individual message, that message is returned undecrypted.

Examples

use matrix_sdk::{room::MessagesOptions, Client};

let request = MessagesOptions::backward("t47429-4392820_219380_26003_2265");

let mut client = Client::new(homeserver).await.unwrap();
let room = client
   .get_joined_room(room_id!("!roomid:example.com"))
   .unwrap();
assert!(room.messages(request).await.is_ok());

pub async fn event(&self, event_id: &EventId) -> Result<RoomEvent>

Fetch the event with the given EventId in this room.

pub async fn sync_members(&self) -> Result<Option<MembersResponse>>

Sync the member list with the server.

This method will de-duplicate requests if it is called multiple times in quick succession, in that case the return value will be None.

pub async fn active_members(&self) -> Result<Vec<RoomMember>>

Get active members for this room, includes invited, joined members.

Note: This method will fetch the members from the homeserver if the member list isn’t synchronized due to member lazy loading. Because of that, it might panic if it isn’t run on a tokio thread.

Use active_members_no_sync() if you want a method that doesn’t do any requests.

pub async fn active_members_no_sync(&self) -> Result<Vec<RoomMember>>

Get active members for this room, includes invited, joined members.

Note: This method will not fetch the members from the homeserver if the member list isn’t synchronized due to member lazy loading. Thus, members could be missing from the list.

Use active_members() if you want to ensure to always get the full member list.

pub async fn joined_members(&self) -> Result<Vec<RoomMember>>

Get all the joined members of this room.

Note: This method will fetch the members from the homeserver if the member list isn’t synchronized due to member lazy loading. Because of that it might panic if it isn’t run on a tokio thread.

Use joined_members_no_sync() if you want a method that doesn’t do any requests.

pub async fn joined_members_no_sync(&self) -> Result<Vec<RoomMember>>

Get all the joined members of this room.

Note: This method will not fetch the members from the homeserver if the member list isn’t synchronized due to member lazy loading. Thus, members could be missing from the list.

Use joined_members() if you want to ensure to always get the full member list.

pub async fn get_member(&self, user_id: &UserId) -> Result<Option<RoomMember>>

Get a specific member of this room.

Note: This method will fetch the members from the homeserver if the member list isn’t synchronized due to member lazy loading. Because of that it might panic if it isn’t run on a tokio thread.

Use get_member_no_sync() if you want a method that doesn’t do any requests.

Arguments

user_id - The ID of the user that should be fetched out of the store.

pub async fn get_member_no_sync(
&self,
user_id: &UserId
) -> Result<Option<RoomMember>>

Get a specific member of this room.

Note: This method will not fetch the members from the homeserver if the member list isn’t synchronized due to member lazy loading. Thus, members could be missing.

Use get_member() if you want to ensure to always have the full member list to chose from.

Arguments

user_id - The ID of the user that should be fetched out of the store.

pub async fn members(&self) -> Result<Vec<RoomMember>>

Get all members for this room, includes invited, joined and left members.

Note: This method will fetch the members from the homeserver if the member list isn’t synchronized due to member lazy loading. Because of that it might panic if it isn’t run on a tokio thread.

Use members_no_sync() if you want a method that doesn’t do any requests.

pub async fn members_no_sync(&self) -> Result<Vec<RoomMember>>

Get all members for this room, includes invited, joined and left members.

Note: This method will not fetch the members from the homeserver if the member list isn’t synchronized due to member lazy loading. Thus, members could be missing.

Use members() if you want to ensure to always get the full member list.

pub async fn get_state_events(
&self,
event_type: EventType
) -> Result<Vec<Raw<AnySyncStateEvent>>>

Get all state events of a given type in this room.

pub async fn get_state_event(
&self,
event_type: EventType,
state_key: &str
) -> Result<Option<Raw<AnySyncStateEvent>>>

Get a specific state event in this room.

pub async fn contains_only_verified_devices(&self) -> Result<bool>

This is supported on crate feature encryption only.

Check if all members of this room are verified and all their devices are verified.

Returns true if all devices in the room are verified, otherwise false.

pub async fn set_tag(
&self,
tag: TagName,
tag_info: TagInfo
) -> HttpResult<Response>

Adds a tag to the room, or updates it if it already exists.

Returns the [create_tag::Response] from the server.

Arguments

tag - The tag to add or update.

tag_info - Information about the tag, generally containing the order parameter.

Example

use matrix_sdk::ruma::events::tag::TagInfo;

if let Some(room) = client.get_joined_room(&room_id) {
    let mut tag_info = TagInfo::new();
    tag_info.order = Some(0.9);
    let user_tag = UserTagName::from_str("u.work")?;

    room.set_tag(TagName::User(user_tag), tag_info ).await?;
}

pub async fn remove_tag(&self, tag: TagName) -> HttpResult<Response>

Removes a tag from the room.

Returns the [delete_tag::Response] from the server.

Arguments

tag - The tag to remove.