Struct matrix_sdk::room::Invited

pub struct Invited { /* fields omitted */ }

A room in the invited state.

This struct contains all methods specific to a Room with type RoomType::Invited. Operations may fail once the underlying Room changes RoomType.

Implementations

impl Invited

pub fn new(client: Client, room: BaseRoom) -> Option<Self>

Create a new room::Invited if the underlying Room has type RoomType::Invited.

Arguments

client - The client used to make requests.

room - The underlying room.

pub async fn reject_invitation(&self) -> Result<()>

Reject the invitation.

pub async fn accept_invitation(&self) -> Result<()>

Accept the invitation.