Struct matrix_sdk::room::Left

pub struct Left { /* fields omitted */ }

A room in the left state.

This struct contains all methodes specific to a Room with type RoomType::Left. Operations may fail once the underlaying Room changes RoomType.

Implementations

impl Left

pub fn new(client: Client, room: BaseRoom) -> Option<Self>

Create a new room::Left if the underlaying Room has type RoomType::Left.

Arguments

client - The client used to make requests.

room - The underlaying room.

pub async fn join(&self) -> Result<()>

Join this room.

pub async fn forget(&self) -> Result<()>

Forget this room.

This communicates to the homeserver that it should forget the room.