Struct matrix_sdk::room::MessagesOptions

#[non_exhaustive]
pub struct MessagesOptions {
    pub from: Option<String>,
    pub to: Option<String>,
    pub dir: Direction,
    pub limit: UInt,
    pub filter: RoomEventFilter,
}

Options for messages.

See that method and https://spec.matrix.org/v1.3/client-server-api/#get_matrixclientv3roomsroomidmessages for details.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

from: Option<String>

The token to start returning events from.

This token can be obtained from a prev_batch token returned for each room from the sync API, or from a start or end token returned by a previous messages call.

If from isn’t provided the homeserver shall return a list of messages from the first or last (per the value of the dir parameter) visible event in the room history for the requesting user.

to: Option<String>

The token to stop returning events at.

This token can be obtained from a prev_batch token returned for each room by the sync API, or from a start or end token returned by a previous messages call.

dir: Direction

The direction to return events in.

limit: UInt

The maximum number of events to return.

Default: 10.

filter: RoomEventFilter

A [RoomEventFilter] to filter returned events with.

Implementations

impl MessagesOptions

pub fn new(dir: Direction) -> Self

Creates MessagesOptions with the given direction.

All other parameters will be defaulted.

pub fn backward() -> Self

Creates MessagesOptions with dir set to Backward.

If no from token is set afterwards, pagination will start at the end of (the accessible part of) the room timeline.

pub fn forward() -> Self

Creates MessagesOptions with dir set to Forward.

If no from token is set afterwards, pagination will start at the beginning of (the accessible part of) the room timeline.

pub fn from<'a>(self, from: impl Into<Option<&'a str>>) -> Self

Creates a new MessagesOptions from self with the from field set to the given value.

Since the field is public, you can also assign to it directly. This method merely acts as a shorthand for that, because it is very common to set this field.