Struct matrix_sdk::ruma::api::client::account::request_registration_token_via_msisdn::v3::Request

#[non_exhaustive]
pub struct Request<'a> {
    pub client_secret: &'a ClientSecret,
    pub country: &'a str,
    pub phone_number: &'a str,
    pub send_attempt: UInt,
    pub next_link: Option<&'a str>,
    pub identity_server_info: Option<IdentityServerInfo<'a>>,
}

This is supported on crate feature api only.

Data for a request to the request_registration_token_via_msisdn API endpoint.

Request a registration token with a phone number.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

client_secret: &'a ClientSecret

Client-generated secret string used to protect this session.

country: &'a str

Two-letter ISO 3166 country code for the phone number.

phone_number: &'a str

Phone number to validate.

send_attempt: UInt

Used to distinguish protocol level retries from requests to re-send the SMS.

next_link: Option<&'a str>

Return URL for identity server to redirect the client back to.

identity_server_info: Option<IdentityServerInfo<'a>>

Optional identity server hostname and access token.

Deprecated since r0.6.0.

Implementations

impl<'a> Request<'a>

pub fn new(
client_secret: &'a ClientSecret,
country: &'a str,
phone_number: &'a str,
send_attempt: UInt
) -> Request<'a>

Creates a new Request with the given client secret, country code, phone number and send-attempt counter.