Trait matrix_sdk::ruma::api::IncomingRequest

pub trait IncomingRequest {
    type EndpointError: EndpointError;
    type OutgoingResponse: OutgoingResponse;

    const METADATA: Metadata;

    fn try_from_http_request<B, S>(
        req: Request<B>, 
        path_args: &[S]
    ) -> Result<Self, FromHttpRequestError>
    where
        B: AsRef<[u8]>,
        S: AsRef<str>;
}

This is supported on crate feature api only.

A request type for a Matrix API endpoint, used for receiving requests.

Associated Types

type EndpointError: EndpointError

A type capturing the error conditions that can be returned in the response.

type OutgoingResponse: OutgoingResponse

Response type to return when the request is successful.

Associated Constants

const METADATA: Metadata

Metadata about the endpoint.

Required methods

fn try_from_http_request<B, S>(
req: Request<B>,
path_args: &[S]
) -> Result<Self, FromHttpRequestError> where
B: AsRef<[u8]>,
S: AsRef<str>,

Tries to turn the given http::Request into this request type, together with the corresponding path arguments.

Note: The strings in path_args need to be percent-decoded.