Trait matrix_sdk::ruma::api::OutgoingResponse

pub trait OutgoingResponse {
    // Required method
    fn try_into_http_response<T>(self) -> Result<Response<T>, IntoHttpError>
       where T: Default + BufMut;
}

A request type for a Matrix API endpoint, used for sending responses.

Required Methods

fn try_into_http_response<T>(self) -> Result<Response<T>, IntoHttpError>where T: Default + BufMut,

Tries to convert this response into an http::Response.

This method should only fail when when invalid header values are specified. It may also fail with a serialization error in case of bugs in Ruma though.