Module matrix_sdk::ruma::common::events

Available on crate feature events only.

(De)serializable types for the events in the Matrix specification. These types are used by other Ruma crates.

All data exchanged over Matrix is expressed as an event. Different event types represent different actions, such as joining a room or sending a message. Events are stored and transmitted as simple JSON structures. While anyone can create a new event type for their own purposes, the Matrix specification defines a number of event types which are considered core to the protocol. This module contains Rust types for all of the event types defined by the specification and facilities for extending the event system for custom event types.

Core event types

This module includes Rust types for all event types in the Matrix specification. To better organize the crate, these types live in separate modules with a hierarchy that matches the reverse domain name notation of the event type. For example, the m.room.message event lives at ruma::events::room::message::RoomMessageEvent. Each type’s module also contains a Rust type for that event type’s content field, and any other supporting types required by the event’s other fields.

Extending Ruma with custom events

For our examples we will start with a simple custom state event. ruma_event specifies the state event’s type and it’s kind.

use ruma_common::events::macros::EventContent;
use serde::{Deserialize, Serialize};

#[derive(Clone, Debug, Deserialize, Serialize, EventContent)]
#[ruma_event(type = "org.example.event", kind = State)]
pub struct ExampleContent {
    field: String,
}

This can be used with events structs, such as passing it into ruma::api::client::state::send_state_event’s Request.

As a more advanced example we create a reaction message event. For this event we will use a OriginalSyncMessageLikeEvent struct but any OriginalMessageLikeEvent struct would work.

use ruma_common::events::{macros::EventContent, OriginalSyncMessageLikeEvent};
use ruma_common::EventId;
use serde::{Deserialize, Serialize};

#[derive(Clone, Debug, Deserialize, Serialize)]
#[serde(tag = "rel_type")]
pub enum RelatesTo {
    #[serde(rename = "m.annotation")]
    Annotation {
        /// The event this reaction relates to.
        event_id: Box<EventId>,
        /// The displayable content of the reaction.
        key: String,
    },

    /// Since this event is not fully specified in the Matrix spec
    /// it may change or types may be added, we are ready!
    #[serde(rename = "m.whatever")]
    Whatever,
}

/// The payload for our reaction event.
#[derive(Clone, Debug, Deserialize, Serialize, EventContent)]
#[ruma_event(type = "m.reaction", kind = MessageLike)]
pub struct ReactionEventContent {
    #[serde(rename = "m.relates_to")]
    pub relates_to: RelatesTo,
}

let json = serde_json::json!({
    "content": {
        "m.relates_to": {
            "event_id": "$xxxx-xxxx",
            "key": "👍",
            "rel_type": "m.annotation"
        }
    },
    "event_id": "$xxxx-xxxx",
    "origin_server_ts": 1,
    "sender": "@someone:example.org",
    "type": "m.reaction",
    "unsigned": {
        "age": 85
    }
});

// The downside of this event is we cannot use it with event enums,
// but could be deserialized from a `Raw<_>` that has failed to deserialize.
matches::assert_matches!(
    serde_json::from_value::<OriginalSyncMessageLikeEvent<ReactionEventContent>>(json),
    Ok(OriginalSyncMessageLikeEvent {
        content: ReactionEventContent {
            relates_to: RelatesTo::Annotation { key, .. },
        },
        ..
    }) if key == "👍"
);

Modules

call

Modules for events in the m.call namespace.

direct

Types for the m.direct event.

dummy

Types for the m.dummy event.

forwarded_room_key

Types for the m.forwarded_room_key event.

fully_read

Types for the m.fully_read event.

ignored_user_list

Types for the m.ignored_user_list event.

key

Modules for events in the m.key namespace.

macros

Re-export of all the derives needed to create your own event types.

policy

Modules for events in the m.policy namespace.

presence

A presence event is represented by a struct with a set content field.

push_rules

Types for the m.push_rules event.

reactionunstable-msc2677

Types for the m.reaction event.

receipt

Types for the m.receipt event.

room

Modules for events in the m.room namespace.

room_key

Types for the m.room_key event.

room_key_request

Types for the m.room_key_request event.

secret

Module for events in the m.secret namespace.

space

Types for the m.space events.

sticker

Types for the m.sticker event.

tag

Types for the m.tag event.

typing

Types for the m.typing event.

Structs

DecryptedMegolmV1Event

The decrypted payload of an m.megolm.v1.aes-sha2 event.

DecryptedOlmV1Event

The decrypted payload of an m.olm.v1.curve25519-aes-sha2 event.

EphemeralRoomEvent

An ephemeral room event.

GlobalAccountDataEvent

A global account data event.

InitialStateEvent

A minimal state event, used for creating a new room.

MessageLikeUnsigned

Extra information about a message event that is not incorporated into the event’s hash.

OlmV1Keys

Public keys used for an m.olm.v1.curve25519-aes-sha2 event.

OriginalMessageLikeEvent

An unredacted message-like event.

OriginalStateEvent

An unredacted state event.

OriginalSyncMessageLikeEvent

An unredacted message-like event without a room_id.

OriginalSyncStateEvent

An unredacted state event without a room_id.

RedactedMessageLikeEvent

A redacted message-like event.

RedactedStateEvent

A redacted state event.

RedactedSyncMessageLikeEvent

A redacted message-like event without a room_id.

RedactedSyncStateEvent

A redacted state event without a room_id.

RedactedUnsigned

Extra information about a redacted event that is not incorporated into the event’s hash.

RoomAccountDataEvent

A room account data event.

StateUnsigned

Extra information about a state event that is not incorporated into the event’s hash.

StrippedStateEvent

A stripped-down state event, used for previews of rooms the user has been invited to.

SyncEphemeralRoomEvent

An ephemeral room event without a room_id.

ToDeviceEvent

An event sent using send-to-device messaging.

Enums

AnyEphemeralRoomEvent

Any ephemeral room event.

AnyEphemeralRoomEventContent

Any ephemeral room event.

AnyGlobalAccountDataEvent

Any global account data event.

AnyGlobalAccountDataEventContent

Any global account data event.

AnyInitialStateEvent

Any state event.

AnyMessageLikeEvent

Any message-like event.

AnyMessageLikeEventContent

Any message-like event.

AnyRoomAccountDataEvent

Any room account data event.

AnyRoomAccountDataEventContent

Any room account data event.

AnyRoomEvent

Any room event.

AnyStateEvent

Any state event.

AnyStateEventContent

Any state event.

AnyStrippedStateEvent

Any state event.

AnySyncEphemeralRoomEvent

Any ephemeral room event.

AnySyncMessageLikeEvent

Any message-like event.

AnySyncRoomEvent

Any sync room event.

AnySyncStateEvent

Any state event.

AnyToDeviceEvent

Any to-device event.

AnyToDeviceEventContent

Any to-device event.

EphemeralRoomEventType

The type of EphemeralRoomEvent this is.

EventKind

The “kind” of an event.

EventTypeDeprecated

The type of Event this is.

GlobalAccountDataEventType

The type of GlobalAccountDataEvent this is.

MessageLikeEvent

A possibly-redacted message-like event.

MessageLikeEventType

The type of MessageLikeEvent this is.

RoomAccountDataEventType

The type of RoomAccountDataEvent this is.

RoomEventType

The type of RoomEvent this is.

StateEvent

A possibly-redacted state event.

StateEventType

The type of StateEvent this is.

SyncMessageLikeEvent

A possibly-redacted message-like event without a room_id.

SyncStateEvent

A possibly-redacted state event without a room_id.

ToDeviceEventType

The type of ToDeviceEvent this is.

Traits

EphemeralRoomEventContent

An alias for EventContent<EventType = EphemeralRoomEventType>.

EventContent

The base trait that all event content types implement.

GlobalAccountDataEventContent

An alias for EventContent<EventType = GlobalAccountDataEventType>.

MessageLikeEventContent

An alias for EventContent<EventType = MessageLikeEventType>.

Redact

Trait to define the behavior of redacting an event.

RedactContent

Trait to define the behavior of redact an event’s content object.

RedactedEventContent

The base trait that all redacted event content types implement.

RedactedMessageLikeEventContent

An alias for EventContent<EventType = MessageLikeEventType> + RedactedEventContent.

RedactedStateEventContent

An alias for EventContent<EventType = StateEventType> + RedactedEventContent.

RoomAccountDataEventContent

An alias for EventContent<EventType = RoomAccountDataEventType>.

StateEventContent

An alias for EventContent<EventType = StateEventType>.

StaticEventContent

Trait for abstracting over event content structs.

ToDeviceEventContent

An alias for EventContent<EventType = ToDeviceEventType>.