Module matrix_sdk::ruma::events::poll

Modules for events in the m.poll namespace (MSC3381).

This module also contains types shared by events in its child namespaces.

Modules

end

Types for the m.poll.end event.

response

Types for the m.poll.response event.

start

Types for the m.poll.start event.

unstable_end

Types for the org.matrix.msc3381.poll.end event, the unstable version of m.poll.end.

unstable_response

Types for the org.matrix.msc3381.poll.response event, the unstable version of m.poll.response.

unstable_start

Types for the org.matrix.msc3381.poll.start event, the unstable version of m.poll.start.

Structs

PollResponseData

The data from a poll response necessary to compile poll results.

Functions

compile_poll_results

Generate the current results with the given poll and responses.

compile_unstable_poll_results

Generate the current results with the given unstable poll and responses.