Struct matrix_sdk::ruma::events::receipt::ReceiptEventContent

pub struct ReceiptEventContent(pub BTreeMap<OwnedEventId, BTreeMap<ReceiptType, BTreeMap<OwnedUserId, Receipt, Global>, Global>, Global>);

The content of an m.receipt event.

A mapping of event ID to a collection of receipts for this event ID. The event ID is the ID of the event being acknowledged and not an ID for the receipt itself.

Tuple Fields
0: BTreeMap<OwnedEventId, BTreeMap<ReceiptType, BTreeMap<OwnedUserId, Receipt, Global>, Global>, Global>
Implementations

impl ReceiptEventContent

pub fn user_receipt( &self, user_id: &UserId, receipt_type: ReceiptType ) -> Option<(&EventId, &Receipt)>

Get the receipt for the given user ID with the given receipt type, if it exists.