Module matrix_sdk::ruma::events::room::member

Types for the m.room.member event.

Structs

Change

A simple representation of a change, containing old and new data.

MembershipDetails

The details of a member event required to calculate a MembershipChange.

RedactedRoomMemberEventContent

A member event that has been redacted.

RedactedThirdPartyInvite

Redacted information about a third party invitation.

RoomMemberEventContent

The content of an m.room.member event.

RoomMemberUnsigned

Extra information about a message event that is not incorporated into the event’s hash.

SignedContent

A block of content which has been signed, which servers can use to verify a third party invitation.

ThirdPartyInvite

Information about a third party invitation.

Enums

MembershipChange

Translation of the membership change in m.room.member event.

MembershipState

The membership state of a user.

Type Aliases

InitialRoomMemberEvent

An m.room.member event for creating a room.

OriginalRoomMemberEvent

An m.room.member event.

OriginalSyncRoomMemberEvent

An m.room.member event from a sync_events response.

PossiblyRedactedRoomMemberEventContent

The possibly redacted form of RoomMemberEventContent.

RedactedRoomMemberEvent

An m.room.member event that has been redacted.

RedactedSyncRoomMemberEvent

An m.room.member event from a sync_events response that has been redacted.

RoomMemberEvent

An m.room.member event.

StrippedRoomMemberEvent

An m.room.member event from an invited room preview.

SyncRoomMemberEvent

An m.room.member event from a sync_events response.