Enum matrix_sdk::ruma::events::secret_storage::key::SecretStorageEncryptionAlgorithm

#[non_exhaustive]
pub enum SecretStorageEncryptionAlgorithm {
    V1AesHmacSha2(SecretStorageV1AesHmacSha2Properties),
}

An algorithm and its properties, used to encrypt a secret.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

V1AesHmacSha2(SecretStorageV1AesHmacSha2Properties)

Encrypted using the m.secret_storage.v1.aes-hmac-sha2 algorithm.

Secrets using this method are encrypted using AES-CTR-256 and authenticated using HMAC-SHA-256.

Implementations

impl SecretStorageEncryptionAlgorithm

pub fn algorithm(&self) -> &str

The algorithm string.

pub fn properties(&self) -> Cow<'_, Map<String, Value>>

The algorithm-specific properties.

The returned JSON object won’t contain the algorithm field, use Self::algorithm() to access it.

Prefer to use the public variants of SecretStorageEncryptionAlgorithm where possible; this method is meant to be used for custom algorithms only.