Struct matrix_sdk::ruma::OwnedTransactionId

pub struct OwnedTransactionId { /* private fields */ }

Owned variant of TransactionId

The wrapper type for this type is variable, by default it’ll use Box, but you can change that by setting “--cfg=ruma_identifiers_storage=...” using RUSTFLAGS or .cargo/config.toml (under [build] -> rustflags = ["..."]) to the following;

ruma_identifiers_storage="Arc" to use Arc as a wrapper type.