Struct matrix_sdk::sliding_sync::SlidingSync

pub struct SlidingSync { /* private fields */ }

Available on crate feature experimental-sliding-sync only.

The Sliding Sync instance.

It is OK to clone this type as much as you need: cloning it is cheap.

Implementations

impl SlidingSync

pub fn builder(id: String, client: Client) -> Result<SlidingSyncBuilder, Error>

Create a new SlidingSyncBuilder.

pub fn subscribe_to_room( &self, room_id: OwnedRoomId, settings: Option<RoomSubscription> )

Subscribe to a given room.

If the associated Room exists, it will be marked as members are missing, so that it ensures to re-fetch all members.

pub fn unsubscribe_from_room(&self, room_id: OwnedRoomId)

Unsubscribe from a given room.

pub async fn get_room(&self, room_id: &RoomId) -> Option<SlidingSyncRoom>

Lookup a specific room

pub fn get_number_of_rooms(&self) -> usize

Check the number of rooms.

pub async fn on_list<Function, FunctionOutput, R>( &self, list_name: &str, function: Function ) -> Option<R>where Function: FnOnce(&SlidingSyncList) -> FunctionOutput, FunctionOutput: Future<Output = R>,

Find a list by its name, and do something on it if it exists.

pub async fn add_list( &self, list_builder: SlidingSyncListBuilder ) -> Result<Option<SlidingSyncList>>

Add the list to the list of lists.

As lists need to have a unique .name, if a list with the same name is found the new list will replace the old one and the return it or None.

pub async fn add_cached_list( &self, list_builder: SlidingSyncListBuilder ) -> Result<Option<SlidingSyncList>>

Add a list that will be cached and reloaded from the cache.

This will raise an error if a storage key was not set, or if there was a I/O error reading from the cache.

The rest of the semantics is the same as Self::add_list.

pub async fn get_rooms<I: Iterator<Item = OwnedRoomId>>( &self, room_ids: I ) -> Vec<Option<SlidingSyncRoom>>

Lookup a set of rooms

pub async fn get_all_rooms(&self) -> Vec<SlidingSyncRoom>

Get all rooms.

pub fn sync(&self) -> impl Stream<Item = Result<UpdateSummary, Error>> + '_

Create a new Sliding Sync sync-loop.

This method returns a Stream, which will send requests and will handle responses automatically. Lists and rooms are updated automatically.

This function returns Ok(…) if everything went well, otherwise it will return Err(…). An Err will always lead to the Stream termination.

pub fn stop_sync(&self) -> Result<()>

Force to stop the sync-loop (Self::sync) if it’s running.

Usually, dropping the Stream returned by Self::sync should be enough to “stop” it, but depending of how this Stream is used, it might not be obvious to drop it immediately (thinking of using this API over FFI; the foreign-language might not be able to drop a value immediately). Thus, calling this method will ensure that the sync-loop stops gracefully and as soon as it returns.

impl SlidingSync

pub fn pos(&self) -> Option<String>

Available on crate feature testing only.

Get a copy of the pos value.

pub fn set_pos(&self, new_pos: String)

Available on crate feature testing only.

Set a new value for pos.

pub fn sliding_sync_proxy(&self) -> Option<Url>

Available on crate feature testing only.

Get the URL to Sliding Sync.

pub async fn force_cache_to_storage( &self, to_device_token: Option<String> ) -> Result<()>

Available on crate feature testing only.

Force caching the current sliding sync to storage.

pub fn extensions_config(&self) -> ExtensionsConfig

Available on crate feature testing only.

Read the static extension configuration for this Sliding Sync.

Note: this is not the next content of the sticky parameters, but rightly the static configuration that was set during creation of this Sliding Sync.