Struct matrix_sdk::sliding_sync::SlidingSyncList

pub struct SlidingSyncList { /* private fields */ }

Available on crate feature experimental-sliding-sync only.

Holding a specific filtered list within the concept of sliding sync.

It is OK to clone this type as much as you need: cloning it is cheap.

Implementations

impl SlidingSyncList

pub fn builder(name: impl Into<String>) -> SlidingSyncListBuilder

Create a new SlidingSyncListBuilder with the given name.

pub fn name(&self) -> &str

Get the name of the list.

pub fn set_sync_mode<M>(&self, sync_mode: M)where M: Into<SlidingSyncMode>,

Change the sync-mode.

It is sometimes necessary to change the sync-mode of a list on-the-fly.

This will change the sync-mode but also the request generator. A new request generator is generated. Since requests are calculated based on the request generator, changing the sync-mode is equivalent to “resetting” the list. It’s actually not calling Self::reset, which means that the state is not reset purposely. The ranges and the state will be updated when the next request will be sent and a response will be received. The maximum number of rooms won’t change.

pub fn state(&self) -> SlidingSyncListLoadingState

Get the current state.

pub fn state_stream( &self ) -> (SlidingSyncListLoadingState, impl Stream<Item = SlidingSyncListLoadingState>)

Get a stream of state updates.

If this list has been reloaded from a cache, the initial value read from the cache will be published.

There’s no guarantee of ordering between items emitted by this stream and those emitted by other streams exposed on this structure.

The first part of the returned tuple is the actual loading state, and the second part is the Stream to receive updates.

pub fn timeline_limit(&self) -> Option<Bound>

Get the timeline limit.

pub fn set_timeline_limit(&self, timeline: Option<Bound>)

Set timeline limit.

pub fn room_list<R>(&self) -> Vec<R>where R: for<'a> From<&'a RoomListEntry>,

Get the current room list.

pub fn room_list_stream( &self ) -> (Vector<RoomListEntry>, impl Stream<Item = VectorDiff<RoomListEntry>>)

Get a stream of room list.

If this list has been reloaded from a cache, the initial value read from the cache will be published.

There’s no guarantee of ordering between items emitted by this stream and those emitted by other streams exposed on this structure.

The first part of the returned tuple is the actual room list entries, and the second part is the Stream to receive updates on the room list entries.

pub fn room_list_filtered_stream<F>( &self, filter: F ) -> (Vector<RoomListEntry>, FilterVectorSubscriber<RoomListEntry, Box<dyn Fn(&RoomListEntry) -> bool + Sync + Send>>)where F: Fn(&RoomListEntry) -> bool + Sync + Send + 'static,

Get a stream of room list, but filtered.

It’s similar to Self::room_list_stream but the room list is filtered by filter.

pub fn maximum_number_of_rooms(&self) -> Option<u32>

Get the maximum number of rooms. See Self::maximum_number_of_rooms to learn more.

pub fn maximum_number_of_rooms_stream(&self) -> impl Stream<Item = Option<u32>>

Get a stream of rooms count.

If this list has been reloaded from a cache, the initial value is published too.

There’s no guarantee of ordering between items emitted by this stream and those emitted by other streams exposed on this structure.

pub fn get_room_id(&self, index: usize) -> Option<OwnedRoomId>

Return the room_id at the given index.

pub fn maybe_commit_sticky(&mut self, txn_id: &TransactionId)

Commit the set of sticky parameters for this list.

pub fn invalidate_sticky_data(&self)

Manually invalidate the sticky data, so the sticky parameters are re-sent next time.

impl SlidingSyncList

pub fn sync_mode(&self) -> SlidingSyncMode

Available on crate feature testing only.

Get the sync-mode.