Struct matrix_sdk::sliding_sync::SlidingSyncRoom

pub struct SlidingSyncRoom { /* private fields */ }

Available on crate feature experimental-sliding-sync only.

A Sliding Sync Room.

It contains some information about a specific room, along with a queue of events for the timeline.

It is OK to clone this type as much as you need: cloning it is cheap, and shallow. All clones of the same value are sharing the same state.

Implementations

impl SlidingSyncRoom

pub fn room_id(&self) -> &RoomId

Get the room ID of this SlidingSyncRoom.

pub fn name(&self) -> Option<String>

This rooms name as calculated by the server, if any

pub fn is_dm(&self) -> Option<bool>

Is this a direct message?

pub fn is_initial_response(&self) -> Option<bool>

Was this an initial response?

pub fn has_unread_notifications(&self) -> bool

Is there any unread notifications?

pub fn unread_notifications(&self) -> UnreadNotificationsCount

Get unread notifications.

pub fn required_state(&self) -> Vec<Raw<AnySyncStateEvent>>

Get the required state.

pub fn prev_batch(&self) -> Option<String>

Get the token for back-pagination.

pub fn timeline_queue(&self) -> Vector<SyncTimelineEvent>

Get a copy of the cached timeline events.

Note: This API only exists temporarily, it will be removed in the future.

pub fn client(&self) -> Client

Get a clone of the associated client.