Module matrix_sdk::store

Functions and types to initialize a store.

The re-exports present here depend on the store-related features that are enabled:

sled-state-store provides a StateStore, while sled-crypto-store provides also a CryptoStore for encryption data. This is the default persistent store implementation for non-WebAssembly.

indexeddb_store provides both a StateStore and a CryptoStore if encryption is also enabled. This is the default persistent store implementation for WebAssembly.

Both options provide a make_store_config convenience method to create a StoreConfig for ClientBuilder::store_config().

Structs

CryptoStore

A sled based cryptostore.

StateStore

Enums

OpenStoreError

All the errors that can occur when opening a sled store.

Functions

make_store_configsled-state-store or sled-crypto-store

Create a StoreConfig with an opened sled StateStore that uses the given path and passphrase. If encryption is enabled, a CryptoStore with the same parameters is also opened.