Struct matrix_sdk::store::StateStore

pub struct StateStore { /* private fields */ }

Implementations

impl SledStore

pub fn open() -> Result<SledStore, StoreError>

pub fn open_with_passphrase(
path: impl AsRef<Path>,
passphrase: &str
) -> Result<SledStore, StoreError>

pub fn open_with_path(path: impl AsRef<Path>) -> Result<SledStore, StoreError>

pub fn open_crypto_store(&self) -> Result<SledStore, OpenStoreError>

Available on crate feature crypto-store only.

Open a CryptoStore that uses the same database as this store.

The given passphrase will be used to encrypt private data.

pub async fn save_filter(
&'_ self,
filter_name: &'_ str,
filter_id: &'_ str
) -> Result<(), SledStoreError>

pub async fn get_filter(
&'_ self,
filter_name: &'_ str
) -> Result<Option<String>, SledStoreError>

pub async fn get_sync_token(&'_ self) -> Result<Option<String>, SledStoreError>

pub async fn save_changes(
&'_ self,
changes: &'_ StateChanges
) -> Result<(), SledStoreError>

pub async fn get_presence_event(
&'_ self,
user_id: &'_ UserId
) -> Result<Option<Raw<PresenceEvent>>, SledStoreError>

pub async fn get_state_event(
&'_ self,
room_id: &'_ RoomId,
event_type: StateEventType,
state_key: &'_ str
) -> Result<Option<Raw<AnySyncStateEvent>>, SledStoreError>

pub async fn get_state_events(
&'_ self,
room_id: &'_ RoomId,
event_type: StateEventType
) -> Result<Vec<Raw<AnySyncStateEvent>, Global>, SledStoreError>

pub async fn get_profile(
&'_ self,
room_id: &'_ RoomId,
user_id: &'_ UserId
) -> Result<Option<MinimalStateEvent<RoomMemberEventContent>>, SledStoreError>

pub async fn get_member_event(
&'_ self,
room_id: &'_ RoomId,
state_key: &'_ UserId
) -> Result<Option<MemberEvent>, SledStoreError>

pub async fn get_user_ids_stream(
&'_ self,
room_id: &'_ RoomId
) -> Result<impl Stream<Item = Result<OwnedUserId, StoreError>>, StoreError>

pub async fn get_stripped_user_ids_stream(
&'_ self,
room_id: &'_ RoomId
) -> Result<impl Stream<Item = Result<OwnedUserId, StoreError>>, StoreError>

pub async fn get_invited_user_ids(
&'_ self,
room_id: &'_ RoomId
) -> Result<impl Stream<Item = Result<OwnedUserId, StoreError>>, StoreError>

pub async fn get_joined_user_ids(
&'_ self,
room_id: &'_ RoomId
) -> Result<impl Stream<Item = Result<OwnedUserId, StoreError>>, StoreError>

pub async fn get_stripped_invited_user_ids(
&'_ self,
room_id: &'_ RoomId
) -> Result<impl Stream<Item = Result<OwnedUserId, StoreError>>, StoreError>

pub async fn get_stripped_joined_user_ids(
&'_ self,
room_id: &'_ RoomId
) -> Result<impl Stream<Item = Result<OwnedUserId, StoreError>>, StoreError>

pub async fn get_room_infos(
&'_ self
) -> Result<impl Stream<Item = Result<RoomInfo, SledStoreError>>, SledStoreError>

pub async fn get_stripped_room_infos(
&'_ self
) -> Result<impl Stream<Item = Result<RoomInfo, SledStoreError>>, SledStoreError>

pub async fn get_users_with_display_name(
&'_ self,
room_id: &'_ RoomId,
display_name: &'_ str
) -> Result<BTreeSet<OwnedUserId>, SledStoreError>

pub async fn get_account_data_event(
&'_ self,
event_type: GlobalAccountDataEventType
) -> Result<Option<Raw<AnyGlobalAccountDataEvent>>, SledStoreError>

pub async fn get_room_account_data_event(
&'_ self,
room_id: &'_ RoomId,
event_type: RoomAccountDataEventType
) -> Result<Option<Raw<AnyRoomAccountDataEvent>>, SledStoreError>