Struct matrix_sdk::BaseRoom

pub struct BaseRoom { /* private fields */ }

The underlying room data structure collecting state for joined, left and invited rooms.

Implementations

impl Room

pub fn room_id(&self) -> &RoomId

Get the unique room id of the room.

pub fn own_user_id(&self) -> &UserId

Get our own user id.

pub fn room_type(&self) -> RoomType

Get the type of the room.

pub fn is_space(&self) -> bool

Whether this room’s RoomType is m.space.

pub fn unread_notification_counts(&self) -> UnreadNotificationsCount

Get the unread notification counts.

pub fn are_members_synced(&self) -> bool

Check if the room has it’s members fully synced.

Members might be missing if lazy member loading was enabled for the sync.

Returns true if no members are missing, false otherwise.

pub fn last_prev_batch(&self) -> Option<String>

Get the prev_batch token that was received from the last sync. May be None if the last sync contained the full room history.

pub fn avatar_url(&self) -> Option<Box<MxcUri, Global>>

Get the avatar url of this room.

pub fn canonical_alias(&self) -> Option<Box<RoomAliasId, Global>>

Get the canonical alias of this room.

pub fn create_content(&self) -> Option<RoomCreateEventContent>

Get the m.room.create content of this room.

This usually isn’t optional but some servers might not send an m.room.create event as the first event for a given room, thus this can be optional.

pub fn is_direct(&self) -> bool

Is this room considered a direct message.

pub fn direct_target(&self) -> Option<Box<UserId, Global>>

If this room is a direct message, get the member that we’re sharing the room with.

Note: The member list might have been modified in the meantime and the target might not even be in the room anymore. This setting should only be considered as guidance.

pub fn is_encrypted(&self) -> bool

Is the room encrypted.

pub fn encryption_settings(&self) -> Option<RoomEncryptionEventContent>

Get the m.room.encryption content that enabled end to end encryption in the room.

pub fn guest_access(&self) -> GuestAccess

Get the guest access policy of this room.

pub fn history_visibility(&self) -> HistoryVisibility

Get the history visibility policy of this room.

pub fn is_public(&self) -> bool

Is the room considered to be public.

pub fn join_rule(&self) -> JoinRule

Get the join rule policy of this room.

pub fn max_power_level(&self) -> i64

Get the maximum power level that this room contains.

This is useful if one wishes to normalize the power levels, e.g. from 0-100 where 100 would be the max power level.

pub fn name(&self) -> Option<String>

Get the m.room.name of this room.

pub fn is_tombstoned(&self) -> bool

Has the room been tombstoned.

pub fn tombstone(&self) -> Option<RoomTombstoneEventContent>

Get the m.room.tombstone content of this room if there is one.

pub fn topic(&self) -> Option<String>

Get the topic of the room.

pub async fn display_name(&'_ self) -> Result<String, StoreError>

Calculate the canonical display name of the room, taking into account its name, aliases and members.

The display name is calculated according to this algorithm.

pub async fn joined_user_ids(
&'_ self
) -> Result<Vec<Box<UserId, Global>, Global>, StoreError>

Get the list of users ids that are considered to be joined members of this room.

pub async fn members(&'_ self) -> Result<Vec<RoomMember, Global>, StoreError>

Get the all RoomMembers of this room that are known to the store.

pub async fn joined_members(
&'_ self
) -> Result<Vec<RoomMember, Global>, StoreError>

Get the list of RoomMembers that are considered to be joined members of this room.

pub async fn active_members(
&'_ self
) -> Result<Vec<RoomMember, Global>, StoreError>

Get the list of RoomMembers that are considered to be joined or invited members of this room.

pub fn clone_info(&self) -> RoomInfo

Clone the inner RoomInfo

pub fn update_summary(&self, summary: RoomInfo)

Update the summary with given RoomInfo

pub async fn get_member(
&'_ self,
user_id: &'_ UserId
) -> Result<Option<RoomMember>, StoreError>

Get the RoomMember with the given user_id.

Returns None if the member was never part of this room, otherwise return a RoomMember that can be in a joined, invited, left, banned state.

pub async fn tags(
&'_ self
) -> Result<Option<BTreeMap<TagName, TagInfo>>, StoreError>

Get the Tags for this room.

pub async fn user_read_receipt(
&'_ self,
user_id: &'_ UserId
) -> Result<Option<(Box<EventId, Global>, Receipt)>, StoreError>

Get the read receipt as a EventId and Receipt tuple for the given user_id in this room.

pub async fn event_read_receipts(
&'_ self,
event_id: &'_ EventId
) -> Result<Vec<(Box<UserId, Global>, Receipt), Global>, StoreError>

Get the read receipts as a list of UserId and Receipt tuples for the given event_id in this room.

pub async fn timeline(
&'_ self
) -> Result<(impl Stream<Item = SyncRoomEvent>, impl Stream<Item = Result<SyncRoomEvent, TimelineStreamError>>), StoreError>

Get two stream into the timeline. First one is forward in time and the second one is backward in time.

pub async fn timeline_forward(
&'_ self
) -> Result<impl Stream<Item = SyncRoomEvent>, StoreError>

Create a stream that returns all events of the room’s timeline forward in time.

If you need also a backward stream you should use timeline

pub async fn timeline_backward(
&'_ self
) -> Result<impl Stream<Item = Result<SyncRoomEvent, TimelineStreamError>>, StoreError>

Create a stream that returns all events of the room’s timeline backward in time.

If you need also a forward stream you should use timeline

pub async fn add_timeline_slice(&'_ self, timeline: &'_ TimelineSlice)

Add a new timeline slice to the timeline streams.