Struct matrix_sdk::BaseRoomMember

pub struct BaseRoomMember { /* private fields */ }

A member of a room.

Implementations

impl RoomMember

pub fn user_id(&self) -> &UserId

Get the unique user id of this member.

pub fn event(&self) -> &Arc<MemberEvent>

Get the original member event

pub fn display_name(&self) -> Option<&str>

Get the display name of the member if there is one.

pub fn name(&self) -> &str

Get the name of the member.

This returns either the display name or the local part of the user id if the member didn’t set a display name.

pub fn avatar_url(&self) -> Option<&MxcUri>

Get the avatar url of the member, if there is one.

pub fn normalized_power_level(&self) -> i64

Get the normalized power level of this member.

The normalized power level depends on the maximum power level that can be found in a certain room, it’s always in the range of 0-100.

pub fn power_level(&self) -> i64

Get the power level of this member.

pub fn name_ambiguous(&self) -> bool

Is the name that the member uses ambiguous in the room.

A name is considered to be ambiguous if at least one other member shares the same name.

pub fn membership(&self) -> &MembershipState

Get the membership state of this member.