Trait matrix_sdk::AsyncTraitDeps

pub trait AsyncTraitDeps: Debug + SendOutsideWasm + SyncOutsideWasm { }

Super trait that is used for our store traits, this trait will differ if it’s used on WASM. WASM targets will not require Send and Sync to have implemented, while other targets will.