Module matrix_sdk::deserialized_responses

SDK-specific variations of response types from Ruma.

Structs

AmbiguityChange

A change in ambiguity of room members that an m.room.member event triggers.

AmbiguityChanges

Collection of ambiguity changes that room member events trigger.

EncryptionInfo

Struct containing information on how an event was decrypted.

MembersResponse

A deserialized response for the rooms members API call.

SyncTimelineEvent

A customized version of a room event coming from a sync that holds optional encryption info.

TimelineEvent

Enums

AlgorithmInfo

The algorithm specific information of a decrypted event.

AnySyncOrStrippedState

Wrapper around both versions of any state event.

DeviceLinkProblem

The sub-enum containing detailed information on why we were not able to link a message back to a device.

RawAnySyncOrStrippedState

Wrapper around both versions of any raw state event.

RawAnySyncOrStrippedTimelineEvent

Wrapper around both versions of any event received via sync.

RawSyncOrStrippedState

Wrapper around both versions of a raw state event.

ShieldState

Recommended decorations for decrypted messages, representing the message’s authenticity properties.

SyncOrStrippedState

Wrapper around both versions of a state event.

VerificationLevel

The sub-enum containing detailed information on why a message is considered to be unverified.

VerificationState

Represents the state of verification for a decrypted message sent by a device.

Type Aliases

MemberEvent

Wrapper around both MemberEvent-Types

RawMemberEvent

Raw version of MemberEvent.