Module matrix_sdk::encryption::verification

Interactive verification for E2EE capable users and devices in Matrix.

The SDK supports interactive verification of devices and users, this module contains types that model and support different verification flows.

A verification flow usually starts its life as a VerificationRequest, the request can then be accepted, or it needs to be accepted by the other side of the verification flow.

Once both sides have agreed to pereform the verification, and the VerificationRequest::is_ready() method returns true, the verification can transition into one of the supported verification flows:

SasVerification - Interactive verification using a short authentication string.

QrVerification - Interactive verification using QR codes.

Structs

AcceptSettings

Customize the accept-reply for a verification process

CancelInfo

Information about the cancellation of a verification request or verification flow.

Emoji

An emoji that is used for interactive verification using a short auth string.

QrVerification

An object controlling QR code style key verification flows.

SasVerification

An object controlling the short auth string verification flow.

VerificationRequest

An object controlling the interactive verification flow.

Enums

QrVerificationData

An enum representing the different modes a QR verification can be in.

ScanError

An error for the different failure modes that can happen during the validation of a scanned QR code.

Verification

An enum over the different verification types the SDK supports.