Struct matrix_sdk::encryption::verification::SasVerification

pub struct SasVerification { /* private fields */ }

Available on crate feature encryption only.

An object controlling the short auth string verification flow.

Implementations

impl SasVerification

pub async fn accept(&self) -> Result<()>

Accept the interactive verification flow.

pub async fn accept_with_settings(&self, settings: AcceptSettings) -> Result<()>

Accept the interactive verification flow with specific settings.

Arguments

settings - specific customizations to the verification flow.

Examples

use matrix_sdk::{
    encryption::verification::{SasVerification, AcceptSettings},
    ruma::events::key::verification::ShortAuthenticationString,
};

let sas = client
    .encryption()
    .get_verification(&user_id, flow_id)
    .await
    .and_then(|v| v.sas());

if let Some(sas) = sas {
    let only_decimal = AcceptSettings::with_allowed_methods(
        vec![ShortAuthenticationString::Decimal]
    );

    sas.accept_with_settings(only_decimal).await?;
}

pub async fn confirm(&self) -> Result<()>

Confirm that the short auth strings match on both sides.

pub async fn mismatch(&self) -> Result<()>

Cancel the interactive verification flow because the short auth strings didn’t match on both sides.

pub async fn cancel(&self) -> Result<()>

Cancel the interactive verification flow.

pub fn emoji(&self) -> Option<[Emoji; 7]>

Get the emoji version of the short auth string.

Examples

use matrix_sdk::{
    encryption::verification::{SasVerification, AcceptSettings},
    ruma::events::key::verification::ShortAuthenticationString,
};

let sas_verification = client
    .encryption()
    .get_verification(&user_id, flow_id)
    .await
    .and_then(|v| v.sas());

if let Some(emojis) = sas_verification.and_then(|s| s.emoji()) {
    let emoji_string = emojis
        .iter()
        .map(|e| format!("{:^12}", e.symbol))
        .collect::<Vec<_>>()
        .join("");

    let description = emojis
        .iter()
        .map(|e| format!("{:^12}", e.description))
        .collect::<Vec<_>>()
        .join("");

    println!("Do the emojis match?\n{}\n{}", emoji_string, description);
}

pub fn decimals(&self) -> Option<(u16, u16, u16)>

Get the decimal version of the short auth string.

pub fn supports_emoji(&self) -> bool

Does this verification flow support emoji for the short authentication string.

pub fn is_done(&self) -> bool

Is the verification process done.

pub fn can_be_presented(&self) -> bool

Are we in a state where we can show the short auth string.

pub fn we_started(&self) -> bool

Did we initiate the verification flow.

pub fn cancel_info(&self) -> Option<CancelInfo>

Get info about the cancellation if the verification flow has been cancelled.

pub fn is_cancelled(&self) -> bool

Is the verification process canceled.

pub fn other_device(&self) -> &ReadOnlyDevice

Get the other users device that we’re verifying.

pub fn started_from_request(&self) -> bool

Did this verification flow start from a verification request.

pub fn is_self_verification(&self) -> bool

Is this a verification that is veryfying one of our own devices.

pub fn own_user_id(&self) -> &UserId

Get our own user id.

pub fn other_user_id(&self) -> &UserId

Get the user id of the other user participating in this verification flow.