Enum matrix_sdk::Error

pub enum Error {
    Http(HttpError),
    AuthenticationRequired,
    SerdeJson(JsonError),
    Io(IoError),
    CryptoStoreError(CryptoStoreError),
    OlmError(OlmError),
    MegolmError(MegolmError),
    DecryptorError(DecryptorError),
    StateStore(StoreError),
    Identifier(IdentifierError),
    Url(UrlParseError),
}

Internal representation of errors.

Variants

Http(HttpError)

Error doing an HTTP request.

AuthenticationRequired

Queried endpoint requires authentication but was called on an anonymous client.

SerdeJson(JsonError)

An error de/serializing type for the StateStore

Io(IoError)

An IO error happened.

CryptoStoreError(CryptoStoreError)

This is supported on encryption only.

An error occurred in the crypto store.

OlmError(OlmError)

This is supported on encryption only.

An error occurred during a E2EE operation.

MegolmError(MegolmError)

This is supported on encryption only.

An error occurred during a E2EE group operation.

DecryptorError(DecryptorError)

This is supported on encryption only.

An error occurred during decryption.

StateStore(StoreError)

An error occurred in the state store.

Identifier(IdentifierError)

An error encountered when trying to parse an identifier.

Url(UrlParseError)

An error encountered when trying to parse a url.

Implementations

impl Error

pub fn uiaa_response(&self) -> Option<&UiaaInfo>

Try to destructure the error into an universal interactive auth info.

Some requests require universal interactive auth, doing such a request will always fail the first time with a 401 status code, the response body will contain info how the client can authenticate.

The request will need to be retried, this time containing additional authentication data.

This method is an convenience method to get to the info the server returned on the first, failed request.