Trait matrix_sdk::event_handler::EventHandlerContext

pub trait EventHandlerContext: Sized { }

Context for an event handler.

This trait defines the set of types that may be used as additional arguments in event handler functions after the event itself.