Struct matrix_sdk::room::Joined

pub struct Joined { /* private fields */ }

A room in the joined state.

The JoinedRoom contains all methods specific to a Room with RoomState::Joined. Operations may fail once the underlying Room changes RoomState.

Implementations

impl Joined

pub async fn leave(&self) -> Result<Left>

Leave this room.

pub async fn ban_user( &self, user_id: &UserId, reason: Option<&str> ) -> Result<()>

Ban the user with UserId from this room.

Arguments

user_id - The user to ban with UserId.

reason - The reason for banning this user.

pub async fn kick_user( &self, user_id: &UserId, reason: Option<&str> ) -> Result<()>

Kick a user out of this room.

Arguments

user_id - The UserId of the user that should be kicked out of the room.

reason - Optional reason why the room member is being kicked out.

pub async fn invite_user_by_id(&self, user_id: &UserId) -> Result<()>

Invite the specified user by UserId to this room.

Arguments

user_id - The UserId of the user to invite to the room.

pub async fn invite_user_by_3pid(&self, invite_id: Invite3pid) -> Result<()>

Invite the specified user by third party id to this room.

Arguments

invite_id - A third party id of a user to invite to the room.

pub async fn typing_notice(&self, typing: bool) -> Result<()>

Activate typing notice for this room.

The typing notice remains active for 4s. It can be deactivate at any point by setting typing to false. If this method is called while the typing notice is active nothing will happen. This method can be called on every key stroke, since it will do nothing while typing is active.

Arguments

typing - Whether the user is typing or has stopped typing.

Examples

use std::time::Duration;

use matrix_sdk::ruma::api::client::typing::create_typing_event::v3::Typing;

let room_id = room_id!("!SVkFJHzfwvuaIEawgC:localhost");

if let Some(room) = client.get_joined_room(&room_id) {
    room.typing_notice(true).await?
}

pub async fn send_single_receipt( &self, receipt_type: ReceiptType, thread: ReceiptThread, event_id: OwnedEventId ) -> Result<()>

Send a request to set a single receipt.

Arguments

receipt_type - The type of the receipt to set. Note that it is possible to set the fully-read marker although it is technically not a receipt.

thread - The thread where this receipt should apply, if any. Note that this must be [ReceiptThread::Unthreaded] when sending a [ReceiptType::FullyRead].

event_id - The EventId of the event to set the receipt on.

pub async fn send_multiple_receipts(&self, receipts: Receipts) -> Result<()>

Send a request to set multiple receipts at once.

Arguments

receipts - The Receipts to send.

If receipts is empty, this is a no-op.

pub async fn enable_encryption(&self) -> Result<()>

Enable End-to-end encryption in this room.

This method will be a noop if encryption is already enabled, otherwise sends a m.room.encryption state event to the room. This might fail if you don’t have the appropriate power level to enable end-to-end encryption.

A sync needs to be received to update the local room state. This method will wait for a sync to be received, this might time out if no sync loop is running or if the server is slow.

Examples

let room_id = room_id!("!SVkFJHzfwvuaIEawgC:localhost");

if let Some(room) = client.get_joined_room(&room_id) {
    room.enable_encryption().await?
}

pub async fn sync_up(&self)

Wait for the room to be fully synced.

This method makes sure the room that was returned when joining a room has been echoed back in the sync. Warning: This waits until a sync happens and does not return if no sync is happening! It can also return early when the room is not a joined room anymore!

pub async fn send( &self, content: impl MessageLikeEventContent, txn_id: Option<&TransactionId> ) -> Result<Response>

Send a room message to this room.

Returns the parsed response from the server.

If the encryption feature is enabled this method will transparently encrypt the room message if this room is encrypted.

Note: If you just want to send a custom JSON payload to a room, you can use the Joined::send_raw() method for that.

Arguments

content - The content of the message event.

txn_id - A locally-unique ID describing a message transaction with the homeserver. Unless you’re doing something special, you can pass in None which will create a suitable one for you automatically.

On the sending side, this field is used for re-trying earlier failed transactions. Subsequent messages must never re-use an earlier transaction ID.

On the receiving side, the field is used for recognizing our own messages when they arrive down the sync: the server includes the ID in the MessageLikeUnsigned field transaction_id of the corresponding SyncMessageLikeEvent, but only for the sending device. Other devices will not see it. This is then used to ignore events sent by our own device and/or to implement local echo.

Example

use matrix_sdk::ruma::{
    events::{
        macros::EventContent,
        room::message::{RoomMessageEventContent, TextMessageEventContent},
    },
    uint, MilliSecondsSinceUnixEpoch, TransactionId,
};

let content = RoomMessageEventContent::text_plain("Hello world");
let txn_id = TransactionId::new();

if let Some(room) = client.get_joined_room(&room_id) {
    room.send(content, Some(&txn_id)).await?;
}

// Custom events work too:
#[derive(Clone, Debug, Deserialize, Serialize, EventContent)]
#[ruma_event(type = "org.shiny_new_2fa.token", kind = MessageLike)]
struct TokenEventContent {
    token: String,
    #[serde(rename = "exp")]
    expires_at: MilliSecondsSinceUnixEpoch,
}

let content = TokenEventContent {
    token: generate_token(),
    expires_at: {
        let now = MilliSecondsSinceUnixEpoch::now();
        MilliSecondsSinceUnixEpoch(now.0 + uint!(30_000))
    },
};
let txn_id = TransactionId::new();

if let Some(room) = client.get_joined_room(&room_id) {
    room.send(content, Some(&txn_id)).await?;
}

pub async fn send_raw( &self, content: Value, event_type: &str, txn_id: Option<&TransactionId> ) -> Result<Response>

Send a room message to this room from a json Value.

Returns the parsed response from the server.

If the encryption feature is enabled this method will transparently encrypt the room message if this room is encrypted.

This method is equivalent to the Joined::send() method but allows sending custom JSON payloads, e.g. constructed using the serde_json::json!() macro.

Arguments

content - The content of the event as a json Value.

event_type - The type of the event.

txn_id - A locally-unique ID describing a message transaction with the homeserver. Unless you’re doing something special, you can pass in None which will create a suitable one for you automatically.

On the sending side, this field is used for re-trying earlier failed transactions. Subsequent messages must never re-use an earlier transaction ID.

On the receiving side, the field is used for recognizing our own messages when they arrive down the sync: the server includes the ID in the StateUnsigned field transaction_id of the corresponding SyncMessageLikeEvent, but only for the sending device. Other devices will not see it. This is then used to ignore events sent by our own device and/or to implement local echo.

Example

use serde_json::json;

let content = json!({
    "body": "Hello world",
});

if let Some(room) = client.get_joined_room(&room_id) {
    room.send_raw(content, "m.room.message", None).await?;
}

pub async fn send_attachment( &self, body: &str, content_type: &Mime, data: Vec<u8>, config: AttachmentConfig ) -> Result<Response>

Send an attachment to this room.

This will upload the given data that the reader produces using the upload() method and post an event to the given room. If the room is encrypted and the encryption feature is enabled the upload will be encrypted.

This is a convenience method that calls the upload() and afterwards the send().

Arguments

body - A textual representation of the media that is going to be uploaded. Usually the file name.

content_type - The type of the media, this will be used as the content-type header.

reader - A Reader that will be used to fetch the raw bytes of the media.

config - Metadata and configuration for the attachment.

Examples

let mut image = fs::read("/home/example/my-cat.jpg")?;

if let Some(room) = client.get_joined_room(&room_id) {
    room.send_attachment(
        "My favorite cat",
        &mime::IMAGE_JPEG,
        image,
        AttachmentConfig::new(),
    ).await?;
}

pub async fn update_power_levels( &self, updates: Vec<(&UserId, Int)> ) -> Result<Response>

Update the power levels of a select set of users of this room.

Issue a power_levels state event request to the server, changing the given UserId -> Int levels. May fail if the power_levels aren’t locally known yet or the server rejects the state event update, e.g. because of insufficient permissions. Neither permissions to update nor whether the data might be stale is checked prior to issuing the request.

pub async fn set_name(&self, name: Option<String>) -> Result<Response>

Sets the name of this room.

pub async fn set_room_topic(&self, topic: &str) -> Result<Response>

Sets a new topic for this room.

pub async fn set_avatar_url( &self, url: &MxcUri, info: Option<ImageInfo> ) -> Result<Response>

Sets the new avatar url for this room.

Arguments

avatar_url - The owned matrix uri that represents the avatar

info - The optional image info that can be provided for the avatar

pub async fn remove_avatar(&self) -> Result<Response>

Removes the avatar from the room

pub async fn upload_avatar( &self, mime: &Mime, data: Vec<u8>, info: Option<ImageInfo> ) -> Result<Response>

Uploads a new avatar for this room.

Arguments

mime - The mime type describing the data

data - The data representation of the avatar

info - The optional image info provided for the avatar, the blurhash and the mimetype will always be updated

pub async fn send_state_event( &self, content: impl StateEventContent<StateKey = EmptyStateKey> ) -> Result<Response>

Send a state event with an empty state key to the homeserver.

For state events with a non-empty state key, see send_state_event_for_key.

Returns the parsed response from the server.

Arguments

content - The content of the state event.

Example

use matrix_sdk::ruma::{
    events::{
        macros::EventContent, room::encryption::RoomEncryptionEventContent,
        EmptyStateKey,
    },
    EventEncryptionAlgorithm,
};

let encryption_event_content = RoomEncryptionEventContent::new(
    EventEncryptionAlgorithm::MegolmV1AesSha2,
);
joined_room.send_state_event(encryption_event_content).await?;

// Custom event:
#[derive(Clone, Debug, Deserialize, Serialize, EventContent)]
#[ruma_event(
    type = "org.matrix.msc_9000.xxx",
    kind = State,
    state_key_type = EmptyStateKey,
)]
struct XxxStateEventContent {/* fields... */}

let content: XxxStateEventContent = todo!();
joined_room.send_state_event(content).await?;

pub async fn send_state_event_for_key<C, K>( &self, state_key: &K, content: C ) -> Result<Response>where C: StateEventContent, C::StateKey: Borrow<K>, K: AsRef<str> + ?Sized,

Send a state event to the homeserver.

Returns the parsed response from the server.

Arguments

content - The content of the state event.

state_key - A unique key which defines the overwriting semantics for this piece of room state.

Example

use matrix_sdk::ruma::{
    events::{
        macros::EventContent,
        room::member::{RoomMemberEventContent, MembershipState},
    },
    mxc_uri,
};

let avatar_url = mxc_uri!("mxc://example.org/avatar").to_owned();
let mut content = RoomMemberEventContent::new(MembershipState::Join);
content.avatar_url = Some(avatar_url);

joined_room.send_state_event_for_key(ruma::user_id!("@foo:bar.com"), content).await?;

// Custom event:
#[derive(Clone, Debug, Deserialize, Serialize, EventContent)]
#[ruma_event(type = "org.matrix.msc_9000.xxx", kind = State, state_key_type = String)]
struct XxxStateEventContent { /* fields... */ }

let content: XxxStateEventContent = todo!();
joined_room.send_state_event_for_key("foo", content).await?;

pub async fn send_state_event_raw( &self, content: Value, event_type: &str, state_key: &str ) -> Result<Response>

Send a raw room state event to the homeserver.

Returns the parsed response from the server.

Arguments

content - The raw content of the state event.

event_type - The type of the event that we’re sending out.

state_key - A unique key which defines the overwriting semantics for this piece of room state. This value is often a zero-length string.

Example

use serde_json::json;

let content = json!({
    "avatar_url": "mxc://example.org/SEsfnsuifSDFSSEF",
    "displayname": "Alice Margatroid",
    "membership": "join"
});

if let Some(room) = client.get_joined_room(&room_id) {
    room.send_state_event_raw(content, "m.room.member", "").await?;
}

pub async fn redact( &self, event_id: &EventId, reason: Option<&str>, txn_id: Option<OwnedTransactionId> ) -> HttpResult<Response>

Strips all information out of an event of the room.

Returns the [redact_event::v3::Response] from the server.

This cannot be undone. Users may redact their own events, and any user with a power level greater than or equal to the redact power level of the room may redact events there.

Arguments

event_id - The ID of the event to redact

reason - The reason for the event being redacted.

txn_id - A unique ID that can be attached to this event as its transaction ID. If not given one is created for the message.

Example

use matrix_sdk::ruma::event_id;

if let Some(room) = client.get_joined_room(&room_id) {
    let event_id = event_id!("$xxxxxx:example.org");
    let reason = Some("Indecent material");
    room.redact(&event_id, reason, None).await?;
}