Struct matrix_sdk::room::Joined

pub struct Joined { /* fields omitted */ }

A room in the joined state.

The JoinedRoom contains all methodes specific to a Room with type RoomType::Joined. Operations may fail once the underlaying Room changes RoomType.

Implementations

impl Joined

pub fn new(client: Client, room: BaseRoom) -> Option<Self>

Create a new room::Joined if the underlaying BaseRoom has type RoomType::Joined.

Arguments

client - The client used to make requests.

room - The underlaying room.

pub async fn leave(&self) -> Result<()>

Leave this room.

pub async fn ban_user(
&self,
user_id: &UserId,
reason: Option<&str>
) -> Result<()>

Ban the user with UserId from this room.

Arguments

user_id - The user to ban with UserId.

reason - The reason for banning this user.

pub async fn kick_user(
&self,
user_id: &UserId,
reason: Option<&str>
) -> Result<()>

Kick a user out of this room.

Arguments

user_id - The UserId of the user that should be kicked out of the room.

reason - Optional reason why the room member is being kicked out.

pub async fn invite_user_by_id(&self, user_id: &UserId) -> Result<()>

Invite the specified user by UserId to this room.

Arguments

user_id - The UserId of the user to invite to the room.

pub async fn invite_user_by_3pid(&self, invite_id: Invite3pid<'_>) -> Result<()>

Invite the specified user by third party id to this room.

Arguments

invite_id - A third party id of a user to invite to the room.

pub async fn typing_notice(&self, typing: bool) -> Result<()>

Activate typing notice for this room.

The typing notice remains active for 4s. It can be deactivate at any point by setting typing to false. If this method is called while the typing notice is active nothing will happen. This method can be called on every key stroke, since it will do nothing while typing is active.

Arguments

typing - Whether the user is typing or has stopped typing.

Examples

use std::time::Duration;
use matrix_sdk::api::r0::typing::create_typing_event::Typing;

room
    .typing_notice(true)
    .await
    .expect("Can't get devices from server");

pub async fn read_receipt(&self, event_id: &EventId) -> Result<()>

Send a request to notify this room that the user has read specific event.

Arguments

event_id - The EventId specifies the event to set the read receipt on.

pub async fn read_marker(
&self,
fully_read: &EventId,
read_receipt: Option<&EventId>
) -> Result<()>

Send a request to notify this room that the user has read up to specific event.

Arguments

fully_read - The EventId of the event the user has read to.

read_receipt - An EventId to specify the event to set the read receipt on.

pub async fn send(
&self,
content: impl Into<AnyMessageEventContent>,
txn_id: Option<Uuid>
) -> Result<Response>

Send a room message to this room.

Returns the parsed response from the server.

If the encryption feature is enabled this method will transparently encrypt the room message if this room is encrypted.

Arguments

content - The content of the message event.

txn_id - A unique Uuid that can be attached to a MessageEvent held in its unsigned field as transaction_id. If not given one is created for the message.

Example

use matrix_sdk::events::{
    AnyMessageEventContent,
    room::message::{MessageEventContent, TextMessageEventContent},
};
use matrix_sdk_common::uuid::Uuid;

let content = AnyMessageEventContent::RoomMessage(
    MessageEventContent::text_plain("Hello world")
);

let txn_id = Uuid::new_v4();
room.send(content, Some(txn_id)).await.unwrap();

pub async fn send_attachment<R: Read>(
&self,
body: &str,
content_type: &Mime,
reader: &mut R,
txn_id: Option<Uuid>
) -> Result<Response>

Send an attachment to this room.

This will upload the given data that the reader produces using the upload() method and post an event to the given room. If the room is encrypted and the encryption feature is enabled the upload will be encrypted.

This is a convenience method that calls the Client::upload() and afterwards the send().

Arguments

body - A textual representation of the media that is going to be uploaded. Usually the file name.

content_type - The type of the media, this will be used as the content-type header.

reader - A Reader that will be used to fetch the raw bytes of the media.

txn_id - A unique Uuid that can be attached to a MessageEvent held in its unsigned field as transaction_id. If not given one is created for the message.

Examples

let path = PathBuf::from("/home/example/my-cat.jpg");
let mut image = File::open(path).unwrap();

room.send_attachment("My favorite cat", &mime::IMAGE_JPEG, &mut image, None)
    .await
    .expect("Can't upload my cat.");

pub async fn send_state_event(
&self,
content: impl Into<AnyStateEventContent>,
state_key: &str
) -> Result<Response>

Send a room state event to the homeserver.

Returns the parsed response from the server.

Arguments

room_id - The id of the room that should receive the message.

content - The content of the state event.

state_key - A unique key which defines the overwriting semantics for this piece of room state. This value is often a zero-length string.

Example

use matrix_sdk::{
    events::{
        AnyStateEventContent,
        room::member::{MemberEventContent, MembershipState},
    },
    identifiers::mxc_uri,
};

let avatar_url = mxc_uri!("mxc://example.org/avatar");
let member_event = MemberEventContent {
   avatar_url: Some(avatar_url),
   membership: MembershipState::Join,
   is_direct: None,
   displayname: None,
   third_party_invite: None,
};

let content = AnyStateEventContent::RoomMember(member_event);
room.send_state_event(content, "").await.unwrap();

pub async fn redact(
&self,
event_id: &EventId,
reason: Option<&str>,
txn_id: Option<Uuid>
) -> Result<Response>

Strips all information out of an event of the room.

Returns the redact_event::Response from the server.

This cannot be undone. Users may redact their own events, and any user with a power level greater than or equal to the redact power level of the room may redact events there.

Arguments

event_id - The ID of the event to redact

reason - The reason for the event being redacted.

txn_id - A unique Uuid that can be attached to this event as its transaction ID. If not given one is created for the message.

Example

let event_id = matrix_sdk::identifiers::event_id!("$xxxxxx:example.org");
let reason = Some("Indecent material");
room.redact(&event_id, reason, None).await.unwrap();